Add files via upload
The aim of the project was waste classification by using Deep Neural Networks. The dataset contains waste images recyclable and organic ones. I have built model to predict if the waste may be recyclable or not. In the analysis I have used Convolutional Neural Network (CNN) model with data augumentation and transfer learning to get more accurate predictions and choose the best one for that.